requirements.txt
numpy==1.16.4
tensorflow-gpu==1.12.0
Keras==2.2.4
flask
opencv-python
requests
Pillow
